requirements.txt
pytest==6.1.2
selenium==3.141.0
allure-pytest==2.8.22